+++ staging/operator-registry/internal/declcfg/declcfg_to_model_test.go
