+++ staging/operator-registry/internal/declcfg/load_test.go
