+++ staging/operator-registry/internal/declcfg/model_to_declcfg.go
@@ -3,8 +3,8 @@ package declcfg
 import (
 	"sort"
 
-	"github.com/operator-framework/operator-registry/pkg/model"
-	"github.com/operator-framework/operator-registry/pkg/property"
+	"github.com/operator-framework/operator-registry/internal/model"
+	"github.com/operator-framework/operator-registry/internal/property"
 )
 
 func ConvertFromModel(mpkgs model.Model) DeclarativeConfig {
