+++ staging/operator-registry/internal/declcfg/testdata/invalid/invalid-bundle-json.json
