+++ staging/operator-registry/internal/declcfg/testdata/invalid/invalid-package-json.json
