+++ staging/operator-registry/internal/declcfg/testdata/invalid/not-json.txt
