+++ staging/operator-registry/internal/declcfg/testdata/valid/cockroachdb.json
