+++ staging/operator-registry/internal/declcfg/testdata/valid/etcd.json
