+++ staging/operator-registry/internal/declcfg/testdata/valid/unrecognized-schema.json
