+++ staging/operator-registry/internal/declcfg/write_test.go
