+++ staging/operator-registry/internal/property/errors.go
