+++ staging/operator-registry/internal/property/property.go
