+++ staging/operator-registry/internal/property/property_test.go
