+++ staging/operator-registry/internal/property/scheme.go
