+++ staging/operator-registry/internal/property/scheme_test.go
